# Fix minor performance issue with color handling