--- a/hedgehog/hedgehog.cabal
+++ b/hedgehog/hedgehog.cabal
@@ -78,6 +78,7 @@ library
 
   ghc-options:
     -Wall
+    -Werror
 
   hs-source-dirs:
     src