Turns out `hSupportsANSI` is not the most efficient call:

```
$ cat spec-before.prof | head -20
	Tue Dec 11 20:42 2018 Time and Allocation Profiling Report  (Final)

	   spec +RTS -p -s -t -RTS

	total time  =        2.03 secs   (2026 ticks @ 1000 us, 1 processor)
	total alloc = 3,087,822,320 bytes  (excludes profiling overheads)

COST CENTRE             MODULE                   SRC                                              %time %alloc

>>=                     Hedgehog.Internal.Tree   src/Hedgehog/Internal/Tree.hs:(136,3)-(141,29)    11.4   11.5
hSupportsANSI.isNotDumb System.Console.ANSI.Unix src/System/Console/ANSI/Unix.hs:74:3-65            8.3   20.6
<*>                     Hedgehog.Internal.Gen    src/Hedgehog/Internal/Gen.hs:(542,3)-(543,6)       7.1    7.1
split                   Hedgehog.Internal.Seed   src/Hedgehog/Internal/Seed.hs:(135,1)-(140,39)     4.3    6.0
randomIvalInteger       System.Random            System/Random.hs:(468,1)-(489,76)                  4.0    3.2
runGenT                 Hedgehog.Internal.Gen    src/Hedgehog/Internal/Gen.hs:(233,1)-(234,13)      4.0    0.2
next                    Hedgehog.Internal.Seed   src/Hedgehog/Internal/Seed.hs:(126,1)-(130,17)     3.3    7.0
list.\                  Hedgehog.Internal.Gen    src/Hedgehog/Internal/Gen.hs:(1302,5)-(1306,31)    3.3    2.6
fmap                    Hedgehog.Internal.Tree   src/Hedgehog/Internal/Tree.hs:(107,3)-(108,34)     2.4    2.9
expand                  Hedgehog.Internal.Tree   src/Hedgehog/Internal/Tree.hs:(85,1)-(89,44)       2.2    2.2
>>=.\                   Hedgehog.Internal.Gen    src/Hedgehog/Internal/Gen.hs:(551,7)-(554,27)      2.0    4.0
```

This is somewhat of a problem because as it stands, hedgehog aggressively checks if the same terminal supports color every property run. This PR makes it check ANSI support once before running and never after, resulting in a 10ish% CPU saving on a CPU-intensive (if quite short) test:

```
$ cat spec-after.prof | head -20
	Tue Dec 11 21:07 2018 Time and Allocation Profiling Report  (Final)

	   spec +RTS -p -s -t -RTS

	total time  =        2.03 secs   (2030 ticks @ 1000 us, 1 processor)
	total alloc = 2,275,820,544 bytes  (excludes profiling overheads)

COST CENTRE            MODULE                     SRC                                                 %time %alloc

>>=                    Hedgehog.Internal.Tree     src/Hedgehog/Internal/Tree.hs:(136,3)-(141,29)       11.2   14.5
<*>                    Hedgehog.Internal.Gen      src/Hedgehog/Internal/Gen.hs:(542,3)-(543,6)          7.8    9.0
split                  Hedgehog.Internal.Seed     src/Hedgehog/Internal/Seed.hs:(135,1)-(140,39)        5.4    7.6
randomIvalInteger      System.Random              System/Random.hs:(468,1)-(489,76)                     4.3    4.1
runGenT                Hedgehog.Internal.Gen      src/Hedgehog/Internal/Gen.hs:(233,1)-(234,13)         3.7    0.3
next                   Hedgehog.Internal.Seed     src/Hedgehog/Internal/Seed.hs:(126,1)-(130,17)        3.7    8.9
list.\                 Hedgehog.Internal.Gen      src/Hedgehog/Internal/Gen.hs:(1302,5)-(1306,31)       3.3    3.2
>>=.\                  Hedgehog.Internal.Gen      src/Hedgehog/Internal/Gen.hs:(551,7)-(554,27)         3.0    5.1
expand                 Hedgehog.Internal.Tree     src/Hedgehog/Internal/Tree.hs:(85,1)-(89,44)          2.9    2.8
fmap                   Hedgehog.Internal.Tree     src/Hedgehog/Internal/Tree.hs:(107,3)-(108,34)        2.5    3.7
integral_.\            Hedgehog.Internal.Gen      src/Hedgehog/Internal/Gen.hs:(804,5)-(809,57)         2.4    2.3
```