@jystic 2019-03-10
I like this change in general but I'm finding it a bit hard to review with the `RunnerConfig` removal as well. If you could keep `RunnerConfig` for now and maybe propose removing it in a separate PR I'd be happy to apply the performance change. That way we can all discuss its merits as it really is a different issue.