@stolyaroleh 2019-11-02
I'd like to point out that the performance hit is not that minor. I just stumbled upon this while writing a simple test for binary search. Given that every iteration was doing <50 `Int` comparisons, it was quite surprising to discover that the `-O2` compiled version took more than few milliseconds to run 1000 iterations!

Here's some more numbers I collected.
With `hSupportsANSI`:
```
	total time  =        0.44 secs   (1326 ticks @ 1000 us, 3 processors)
	total alloc = 2,418,786,712 bytes  (excludes profiling overheads)

COST CENTRE           MODULE                             SRC                                                     %time %alloc

hSupportsANSI         System.Console.ANSI.Unix           src/System/Console/ANSI/Unix.hs:(71,1)-(74,65)           71.5   90.3
async                 Control.Concurrent.Async           Control/Concurrent/Async.hs:194:1-35                     12.7    4.9
```
Without:
```
	total time  =        0.05 secs   (153 ticks @ 1000 us, 3 processors)
	total alloc = 101,967,872 bytes  (excludes profiling overheads)

COST CENTRE             MODULE                        SRC                                                      %time %alloc

async                   Control.Concurrent.Async      Control/Concurrent/Async.hs:194:1-35                      47.1   36.8
runGenT                 Hedgehog.Internal.Gen         src/Hedgehog/Internal/Gen.hs:(240,1)-(241,13)              8.5    7.2
evalIO                  Hedgehog.Internal.Property    src/Hedgehog/Internal/Property.hs:(755,1)-(756,71)         7.8    7.4
```
1000 iterations of a simple test allocate 2.4 GB!

I'd like to bump this PR. If @christianlavoie does not mind, I can implement changes @moodmosaic requested either on top of this PR or as a separate one, with just one small commit.