@moodmosaic 2019-11-05
@stolyaroleh, sure, go ahead :+1:

IIRC, the most important thing was to keep the existing stuff in `RunnerConfig`, and so then we can later decide to clean up separately if and when needed, as per https://github.com/hedgehogqa/haskell-hedgehog/pull/240#issuecomment-471242309.