@moodmosaic 2020-01-08
Thanks for all the work @christianlavoie and @stolyaroleh. :+1: :rocket: Closing this as it's resolved via #345. Feel free to reopen or add comments to it, if needed.